Smoke test SIAF in new refactored arrangement
